transpile: Move `impl From<CBinOp> for BinOp` to `c_ast`
